requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic==2.5.0

# MongoDB
pymongo==4.6.1
motor==3.3.2

# Embeddings and AI Models
voyageai==0.2.3
google-generativeai==0.3.2
dashscope==1.14.1

# PDF Processing
PyMuPDF==1.23.8
pdfplumber==0.10.3
camelot-py[cv]==0.11.0
opencv-python==4.8.1.78
ghostscript==0.7

# Optional: LayoutParser (for advanced PDF table/figure detection)
# layoutparser[paddledetection]==0.3.4
# or
# layoutparser==0.3.4
# detectron2 @ git+https://github.com/facebookresearch/detectron2.git

# Video Processing
openai-whisper==20231117
yt-dlp==2023.12.30
moviepy==1.0.3
torch==2.1.2
torchaudio==2.1.2

# Website Extraction
trafilatura==1.6.3
requests==2.31.0

# Text Processing
langchain==0.1.0
langchain-text-splitters==0.0.1

# Data Processing
pandas==2.1.4
numpy==1.26.2
Pillow==10.1.0

# Utilities
tqdm==4.66.1
aiohttp==3.9.1
asyncio==3.4.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
flake8==6.1.0
ipython==8.18.1
jupyter==1.0.0